stk: Add support of Help information
BZ: 2883

According to the 3GPP TS 31.124, Help
information is an optional feature. If help
information is set as supported, then the
proactive command should support Help menu.
If Help menu is supported but not available
then conformance tests cases are considered
as failed(eg: 27.22.4.8.2 sequence 2.1).

This patch adds support for the optional
feature - Help Info.

Change-Id: Idb1c208e01080b0c5d88296b1d0ac4e8046073be
Origin-Change-Id: I96ff1ce86877c02d9b934cc645cbbefd1d5c3205
Origin-change-Id: Ic58feee3ea7abf4d833f3fe616486fee08bccc61
Signed-off-by: Guillaume Lucas <guillaumex.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/75385
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>