[PORT FROM MAIN] Stk: Implement STK busy screen feature
BZ: 2770

If a normal priority DISPLAY TEXT proactive
command is received when the screen is
busy, then the terminal should respond with
general result "Terminal currently unable
to process command" with additional information
set to screen busy. Currently, this support
is missing.

This patch adds the support by rejecting
the DISPLAY TEXT proactive command with
screen busy if the device is in one
of the busy screens.

Screen definitions:
Idle screen - Screen that is seen in the idle
state.
Busy screen - Opposite of idle screen.
For e.g. : {Incall, SMS, Settings, apps list}.

Change-Id: I87e0555e6fc28e7f65267c5b4e1edcd5ee7447ec
Origin-Change-Id: Ib59b948b9e0e61b3c6d24788182733399d9dffa0
Signed-off-by: Javen Ni <javen.ni@borqs.com>
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Signed-off-by: Nizar Haouati <nizar.haouati@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/77552
Reviewed-by: Amegah, AndreX <andrex.amegah@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>