[PORT FROM MAIN]Stk: Add support for Select Item - Default Item
BZ: 61229

As per 3GPP 11.14, if the Select Item is received
with Item Identifier set to one of the items, then
the ME should select/highlight the default item.
Currently, default item is not selected/highlighted.

This patch fixes the issue by highlighting the default
item by setting the background color.

Change-Id: I28ca01b749cf0b7121be4e6e9b8fbdfd257eca45
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/75285
Reviewed-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>