[PORT FROM MAIN] Stk: Fix issue in DISPLAY TEXT-Sustained text
BZ: 61998

3GPP 27.22.4.1.4/1 Display Text - sustained text
is failing.

As per the CAT specification, text shall
be sustained until a subsequent proactive command
is received containing display data. Existing
dialog showing the text is not replaced with
the newly received text message resulting in
conformance case failing.

waitForUserResponse not set correctly and also
Intent.FLAG_ACTIVITY_MULTIPLE_TASK is used
resulting in a starting of a new task even if there
is an existing task.

Fix the issue by setting the waitForUserResponse
to false on response not needed and on screen
not available. Also, Intent.FLAG_ACTIVITY_CLEAR_TASK
is used as one of the flags in creating the intent
for launching the dialog.

Change-Id: I299cb115b58c21c0243c0f9a57e056941b92bfc3
Origin-Change-Id: I173c0c19ecbb6816cee3d84f8815866a5bcc1e6d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/78149
Reviewed-by: Amegah, AndreX <andrex.amegah@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>