Stk: Fix issues in handling launch browser command
BZ: 60448

DO NOT UPSTREAM

3GPP TS 31.124 test sequence 27.22.4.26.2/3
Launch Browser - if not already launched is
failing.

If the proactive command "Launch browser:
if not already launched" is received and the
browser is already launched then the terminal
should send the terminal response with result
code set to LAUNCH_BROWSER_ERROR and additional
information set to STK_BROWSER_UNAVAILABLE.
Currently, this is not the case.

Note: This patch only works if Stock Android browser
is used to perform the above test.

Change-Id: I7a8d9d24efc82804b528f0a6c4e069592c6eed07
Origin-Change-Id: Icf9e95f7663e6a3fd38c25cce64e4330cd612de1
Origin-Change-Id: I245175407bbed9d0e0870c7e91fa27185f0dd4e4
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/79839
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>