Stk: request user confirmation on OPEN CHANNEL cmd
BZ: 52207

Currently, user confirmation not requested
on OPEN CHANNEL proactive command.

With this patch, user confirmation requested.
If the user accepts, then the channel will
be opended. If the user does not accept the
channel set-up, the terminal informs the UICC
using TERMINAL RESPONSE (User did not accept
the proactive command) and the operation is
aborted.

Change-Id: I371bac2f0f08d1c2217bba20d791a090b538007b
Origin-Change-Id: I73149624542b533b45baceac5d0bbf8fd90fdf0a
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/79842
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>