[PORT FROM R41] Stk: Fix issue in HCI connectivity event download
BZ: 76259

UICC reader mode UC works works only one time
after phone boot, on the subsequent retries it fails.

HCI connectivity event download is done with oneShot
set to true. So, this event is removed from the event
list after download. When the event download is done
again, event is not sent to UICC because the event
is not any more in the event list.

Fix the issue by setting the oneShot parameter to
false when downloading the HCI connectivity event.

Change-Id: I5567fd7d207dbdcda4bc79643d31ef3d827a895e
Orig-Change-Id: I5ed6ca3cee8c2481d107d58580906d1b830dc4ca
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/84706
Reviewed-by: Amegah, AndreX <andrex.amegah@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>