stk: add a check for StkService in handleCmd
BZ: 81819

If the mStkService pointer is not yet initialized when we go to the
handleCmd() function we will have a JAVACRASH. This can happen in
rare condition.

This patch fixes the issue by adding a check to the mStkService
pointer before to use it in the handleCmd function. If the pointer
is not yet set we set it by calling the getInstance function.

Change-Id: I52da0c5dda3d9d17a1c21476716435bda838f321
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/88951
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>