stk: add a check for StkServie in handleCmd
BZ: 84536

If the mStkService pointer is not yet initialized when we go to the
handleCmd() funtion we can have a JAVACRASH for some DISPLAY TEXT
operation.

This patch fixes the issue by adding a check to the mStkService
pointer before to use it. If the pointer is not yet set, we can
to set it by calling the getInstance function.

Change-Id: Ic4828610bfdbdd730eb84b8d9c3fc7b0e594e1f7
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/89774
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>