StkAppService: add SET_ACTIVITY_WATCHER permission
BZ: 84370

The StkAppService needs to have the SET_ACTIVITY_WATCHER permission
to use the registerProcessObserver() function. If this is not done
a FATAL EXCEPTION can be raised for the SETUP EVENT LIST proactive
command.

This patch fixes the issue by adding the necessary permission in the
AndroidManifest.xml file of the StkAppService application.

Change-Id: Ic9f584659d76f7c94f55695ee5593fcefa0b7251
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/94797
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>