stk: add the icon in the notification message for setup idle text
BZ: 83268

The 3GPP conformance test 27.22.4.22.2 is reported as failed because
the icon present in the SETUP_IDLE_MODE_TEXT proactive command is not
present in the notification message.

This issue is due to the fact the icon is set in the content view
attached to the notifiation and not directly to the notification.

This path fixes the issue by setting the large icon of the notification
with the icon provided by the proactive command.

Change-Id: I8bf2e79b27733664eff69775b7ed03e32281c9a7
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/95112
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>