[PORT FROM MAIN] Stk: Increase the priority to receive BOOT_COMPLETED
BZ: 92605

STK app service needed to process the proactive command
is not yet started by Android when the test is done.
CatService activates the receival of proactive command
but the received command is queued till the StkAppService
is available. When STK app service is started, it process
the message correctly. Time taken to process the proactive
command has exceeded the conformance test timeout.

Set android:priority to a positive value which controls
the order in which the broadcast receivers are executed
to receive broadcast messages.

Category: Device Enablement
Domain: Telephony-ModemIntel
Severity: Medium
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I03cc523b6d80c3a6a5e4fe4a6486eacf0f86f002
Orig-Change-Id: I82a36607658c127ac4d7de3f26d1bb71e1e76b3b
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/121542
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>