[PORT FROM MAIN] Stk: Increase the priority to receive BOOT_COMPLETED
BZ: 92605

STK app service needed to process the proactive command
is not yet started by Android when the test is done.
CatService activates the receival of proactive command
but the received command is queued till the StkAppService
is available. When STK app service is started, it process
the message correctly. Time taken to process the proactive
command has exceeded the conformance test timeout.

Set android:priority to a positive value which controls
the order in which the broadcast receivers are executed
to receive broadcast messages.

Severity: medium

Change-Id: Id0bb5a0952956d11199110b11de4cf1f1806ba10
Orig-Change-Id: I03cc523b6d80c3a6a5e4fe4a6486eacf0f86f002
Category: device enablement
Domain: Telephony-ModemIntel
Origin: internal
Upstream-Candidate: no,proprietary
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>