[PORT FROM R42S] Stk: Increase the priority to receive BOOT_COMPLETED
BZ: 92605

[DO NOT UPSTREAM]

STK app service needed to process the proactive command
is not yet started by Android when the test is done.
CatService activates the receival of proactive command
but the received command is queued till the StkAppService
is available. When STK app service is started, it process
the message correctly. Time taken to process the proactive
command has exceeded the conformance test timeout.

Set android:priority to a positive value which controls
the order in which the broadcast receivers are executed
to receive broadcast messages.

Change-Id: I82a36607658c127ac4d7de3f26d1bb71e1e76b3b
Orig-Change-Id: I6f17973fb6462f9c6ba3dde19a6bc996c3ff3fd3
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/110011
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>