[PORT FROM R42S] stk: retrieve the CAT Service instance for setup event list processing
BZ: 105070

[DO NOT UPSTREAM]

The 3GPP conformance tests 27.22.7.9.1 and 27.22.7.6.1 failed because
the Cat Service instance is null when the setup event list is
processed.

This is due to a recent change: the StkAppService is now started
before the the Cat Service. Is it so not possible for the StkAppService
to get the instance of the Cat Service during its creation.

This patch updates the code to retrieve the Cat Service instance.
With this one the 3GPP conformance tests are now passed.

Change-Id: I23f72e5833df7261cf053969559fe00beb84d482
Orig-Change-Id: I7e27198395c525057c26488a46612d647b619cce
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/110013
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>