[PORT FROM MAIN] stk: retrieve the CAT Service instance for setup event list processing
BZ: 105070

The 3GPP conformance tests 27.22.7.9.1 and 27.22.7.6.1 failed because
the Cat Service instance is null when the setup event list is
processed.

This is due to a recent change: the StkAppService is now started
before the the Cat Service. Is it so not possible for the StkAppService
to get the instance of the Cat Service during its creation.

This patch updates the code to retrieve the Cat Service instance.
With this one the 3GPP conformance tests are now passed.

Category: Device Enablement
Domain: Telephony-ModemIntel
Severity: Medium
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: Iea0f9488cd8581dab7bf88dc5e652abb77c4114c
Orig-Change-Id: I23f72e5833df7261cf053969559fe00beb84d482
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/121544
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>