[PORT FROM MAIN] stk: retrieve the CAT Service instance for setup event list processing
BZ: 105070

The 3GPP conformance tests 27.22.7.9.1 and 27.22.7.6.1 failed because
the Cat Service instance is null when the setup event list is
processed.

This is due to a recent change: the StkAppService is now started
before the the Cat Service. Is it so not possible for the StkAppService
to get the instance of the Cat Service during its creation.

This patch updates the code to retrieve the Cat Service instance.
With this one the 3GPP conformance tests are now passed.

Category: Device Enablement
Domain: Telephony-ModemIntel
Severity: Medium
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I9b6741ac49c91f9dc97c4fe34e84a30999f9b0a2
Orig-Change-Id: Iea0f9488cd8581dab7bf88dc5e652abb77c4114c
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>