[PORT FROM MAIN] stk: cancel timeout if the STK main menu is called from intent
BZ: 113076

When a STK end session is handled, the StkMenuActivity is
called by an intent to display the STK main menu. It's
possible in this case to have a timer ongoing due to a
previous resume of the StkMenuActivity for a second menu
level. We need to cancel the timer in this case as there
is not time out for the main menu.

This patch fixes the issue by cancelling the timer if the
StkMenuActivity is called from an intent for the main
menu.

Severity: medium

Change-Id: I4c7b86162db1b2d0b4ef77a4fe0f2c9382075f8c
Orig-Change-Id: I1b71a9ec4f2be04973e11e4e212993b8b4a07f2b
Category: device enablement
Domain: Telephony-ModemIntel
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>