[PORT FROM MAIN] stk: cancel timeout if the STK main menu is called from intent
BZ: 113076

When a STK end session is handled, the StkMenuActivity is
called by an intent to display the STK main menu. It's
possible in this case to have a timer ongoing due to a
previous resume of the StkMenuActivity for a second menu
level. We need to cancel the timer in this case as there
is not time out for the main menu.

This patch fixes the issue by cancelling the timer if the
StkMenuActivity is called from an intent for the main
menu.

Category: Device Enablement
Domain: Telephony-ModemIntel
Severity: Medium
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I1b71a9ec4f2be04973e11e4e212993b8b4a07f2b
Orig-Change-Id: Ida7f63b7b09812e8f450cf4dbcee31f1bc6c77e8
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/121545
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>