[PORT FROM R42S] Stk: Fix Launch browser - if not already launched.
BZ: 116550

[DO NOT UPSTREAM]

STK conformance test 31.124.27.22.26.2/3 - launch
browser if not already launched fails as the default
URL is launched even though the browser is in use.

This patch fixes the issue by do not call the
launchBrowser() if the command qualifier field of
the received launch browser proactive command is
set to "launch browser, if not already launched"
and then the browser is in use.

Change-Id: I11dac3e8d131faed5acb27f2e32a5e82c3dedd10
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Reviewed-on: http://android.intel.com:8080/113844
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>