[PORT FROM MAIN] stk: update stk activities for Android 3.0 and higher
BZ: 110864

The current stk activities are not fully compatible
with Android 3.0. Due to this we're not able to see,
for example, the option menu when this one should be
visible.

This patch updates the code to made it compatible with
Android 3.0 and higher.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I90c851e9ac03756730012092e8fba4d3e6c6c630
Orig-Change-Id: I3c1ccb9349efc43186cf8d146d22d1299c35af66
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/124368
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>