[PORT FROM MAIN] stk: update stk activities for Android 3.0 and higher
BZ: 110864

The current stk activities are not fully compatible
with Android 3.0. Due to this we're not able to see,
for example, the option menu when this one should be
visible.

This patch updates the code to made it compatible with
Android 3.0 and higher.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I615f9cd89d33e1e2c1e4d8c1569931d8cdc2318c
Orig-Change-Id: I90c851e9ac03756730012092e8fba4d3e6c6c630
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>