[PORT FROM MAIN] Stk: Fix issues in handling launch browser command
BZ: 60448 111414 116550

3GPP TS 31.124 test sequence 27.22.4.26.2/3
Launch Browser - if not already launched is
fails as the default URL is launched even though
the browser is in use.

If the proactive command "Launch browser:
if not already launched" is received and the
browser is already launched then the terminal
should send the terminal response with result
code set to LAUNCH_BROWSER_ERROR and additional
information set to STK_BROWSER_UNAVAILABLE.
Currently, this is not the case.

Note: This patch only works if Stock Android browser
is used to perform the above test.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I53554e6f43b2fd39c0917bdc80a0dc1fb0357861
Orig-Change-Id: I7a8d9d24efc82804b528f0a6c4e069592c6eed07
Orig-Change-Id: I31a94fc13ebb26eb518670dc2edcd2ffcf115a57
Orig-Change-Id: I11dac3e8d131faed5acb27f2e32a5e82c3dedd10
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/123874
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>