[PORT FROM MAIN] Stk: Fix issues in handling launch browser command
BZ: 60448 111414 116550

3GPP TS 31.124 test sequence 27.22.4.26.2/3
Launch Browser - if not already launched is
fails as the default URL is launched even though
the browser is in use.

If the proactive command "Launch browser:
if not already launched" is received and the
browser is already launched then the terminal
should send the terminal response with result
code set to LAUNCH_BROWSER_ERROR and additional
information set to STK_BROWSER_UNAVAILABLE.
Currently, this is not the case.

Note: This patch only works if Stock Android browser
is used to perform the above test.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I0df14d6ead3d5e1e77376dde67e739d4f504d747
Orig-Change-Id: I53554e6f43b2fd39c0917bdc80a0dc1fb0357861
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>