stk: update message for DISPLAY_TEXT on new intent
BZ: 160450

When the STK application receives two DISPLAY_TEXT command,
the second dialog box is displayed in background and it's
not possible for the user to click one of the buttons.

This patch updates the code of the StkDialogActivity class
to update the message of the first dialog box via the
onNewInstance() function to fix the issue.

Change-Id: I8cef84f0bcf13a7322f15ba1d61b9a97bafd3b2b
Category: aosp improvement
Domain: Telephony-STK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>