stk: add tabulation support for launch browser command
BZ: 142560

As per 3GPP specification 31.124 if browser supports
multiple sessiond/tabs, it's a valid behavior to open the
session in a new activity tab that doesn't interfere with
other sessions for the LAUNCH_BROWSER if not already
launched proative command.

According to the 3GPP specification 102 223, it's also not
necessary to ask for the user confirmation if there is no
alpha identifier provided.

This patch updates the STK application to take those points
into account.

Change-Id: I1d5646e9c6df401c012e5e74c3d36a848856579a
Category: aosp improvement
Domain: Telephony-STK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>