Stk: Fix for conformance test 27.22.4.1.2/1 failure
BZ: 172689

Conformance test case 27.22.4.1.2/1 fails as device
took more than 70seconds to send terminal response.

As per the conformance test case, if there is no response
from user after 60seconds of the command receival, device
should send terminal response with general result set to
"No response from user". Conformance test case fails as the
device takes more than 70seconds to send the terminal response.

As CAT_CMD_ACTION is an ordered background broadcast, it has
to wait for previous ordered broadcasts to be handled first.
Due to this, StkCmdReceiver is not receiving the CAT_CMD_ACTION
broadcast as soon as it is broadcasted by the android telephony
framework. This is causing delay in sending the terminal
response.

Following changes are done to fix the issue
- StkAppService is now started only on BOOT_COMPLETED event.
- Upon BOOT_COMPLETED event, inform framework that app is ready
to process proactive commands and also process pending proactive
commands if there is any. This is possible if proactive command is
received before the StkAppService is started.
- Dynamically register receiver for the broadcasts CAT_CMD_ACTION,
CAT_SESSION_END_ACTION and USER_ACTIVITY_AVAILABLE_ACTION.
- Also, uninstall the STK app if the sim app state is unknown
or absent or not ready.

Change-Id: Ib5c73a548267c6965e368d32a25d3af1bc4410a7
Category: aosp improvement
Domain: Telephony-STK
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>