<Telephony Cert><IMINAN-16663> stk: update stk activities for Android 3.0 and higher
stk: update stk activities for Android 3.0 and higher

The current stk activities are not fully compatible
with Android 3.0. Due to this we're not able to see,
for example, the option menu when this one should be
visible.

This patch update the code to made it compatible with
Android 3.0 and higher

Category: aosp improvement
Domain: Telephony-STK
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16663
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: https://android.intel.com:443/318220

Orig-Change-Id: I3c1ccb9349efc43186cf8d146d22d1299c35af66
Change-Id: Ic200631c2164615bd95f38cca5c11dee464f4c7e
UTP-Id: SMS06581083
Reviewed-on: https://opticm6.rds.intel.com/r/66907
Tested-by: TCloud Jenkins
Reviewed-by: Xin Teng <xin.teng@intel.com>
Reviewed-by: Qiang Rao <qiang.rao@intel.com>
Reviewed-by: Vishwanath Manvi <vishwanath.manvi@intel.com>
Reviewed-by: UTP Service